Delete README.md
This Computer Store Management project is coded in C++programming language. Talking about the system, the user can make a list of order details as records. You just have to type the required order information and press enter to add the information on the record. The details of customers’ orders and their contacts are stored in the database in the form of a text file. This project follows CRUD operations and is very easy to use and understand.

How To Run The Project?
In order to run the project, you must have installed Dev C++ or Code Blocks on your PC